+++ next/src/components/Img/index.tsx
@@ -17,8 +17,6 @@ function Img({
   height?: number;
   title?: string;
 }) {
-  // Next.js image is now handling alt attribute check
-  // if (!alt) console.error('Missing ALT attribute on IMG') // eslint-disable-line no-console
   return <Image className={className} src={src} alt={alt} width={width} height={height} title={title} />
 }
